requirements.txt
# ============================================================================
# Ева 2.0 - Зависимости проекта
# ============================================================================

# ----------------------------------------------------------------------------
# LLM и AI
# ----------------------------------------------------------------------------
openai>=1.0.0              # Официальный SDK OpenAI для работы с LLM (GPT-4o, GPT-5.x и др.)


# ----------------------------------------------------------------------------
# База данных (PostgreSQL)
# ----------------------------------------------------------------------------
psycopg2-binary>=2.9.0     # Адаптер PostgreSQL (бинарная версия, удобно для MVP и локальной разработки)
sqlalchemy>=2.0.0          # ORM для работы с БД (опционально, для более высокоуровневой работы)

# ----------------------------------------------------------------------------
# Docker
# ----------------------------------------------------------------------------
docker>=6.0.0              # Python SDK для управления Docker-контейнерами

# ----------------------------------------------------------------------------
# Анализ и метрики кода
# ----------------------------------------------------------------------------
radon>=6.0.0               # Метрики качества кода (цикломатическая сложность, поддерживаемость)
astor>=0.8.1               # Обратная генерация кода из AST (для code_editor)

# ----------------------------------------------------------------------------
# Проверка кода (линтеры)
# ----------------------------------------------------------------------------
flake8>=6.0.0              # Линтер для проверки стиля кода
pylint>=2.17.0             # Статический анализатор кода (опционально)

# ----------------------------------------------------------------------------
# Тестирование
# ----------------------------------------------------------------------------
pytest>=7.4.0              # Фреймворк для тестирования
pytest-cov>=4.1.0          # Покрытие кода тестами (опционально)

# ----------------------------------------------------------------------------
# Конфигурация и утилиты
# ----------------------------------------------------------------------------
python-dotenv>=1.0.0       # Загрузка переменных окружения из .env файла
pyyaml>=6.0                # Работа с YAML файлами (конфиги, reflection_log)

# ----------------------------------------------------------------------------
# API и интерфейсы (опционально на MVP)
# ----------------------------------------------------------------------------
fastapi>=0.104.0           # Веб-фреймворк для API (если будет eva_api/api.py)
uvicorn[standard]>=0.24.0 # ASGI сервер для FastAPI
click>=8.1.0               # CLI фреймворк (для eva_api/cli.py)

# ----------------------------------------------------------------------------
# Вспомогательные библиотеки
# ----------------------------------------------------------------------------
pydantic>=2.0.0            # Валидация данных (может использоваться в FastAPI и для схем)
rich>=13.0.0               # Красивый вывод в консоль (для CLI и логов)

# ----------------------------------------------------------------------------
# Примечания:
# ----------------------------------------------------------------------------
# - ast - встроенный модуль Python, не требует установки
# - unittest - встроенный модуль Python, не требует установки
# - json - встроенный модуль Python, не требует установки
# - os, sys, pathlib - встроенные модули Python

